+++ src/game/enemy.cpp
@@ -79,27 +79,23 @@ bool Enemy::interact(std::shared_ptr<Object> actor) {
       }
     }
   }
-
-  auto ptr = shared_from_this();
-  if (hp > 0) {
-    auto damage = hero->getDamage(shared_from_this());
-    auto def = R::Z(0, DEF(this));
-    if (damage->damage - def < 0) {
-      damage->damage = 0;
-    } else {
-      damage->damage -= def;
-    }
-    damage->deflected = def;
-    hp -= damage->damage;
-    EnemyTakeDamageEvent e(ptr, damage);
-    eb::EventBus::FireEvent(e);
-  }
-  if (hp <= 0) {
-    passThrough = true;
-    EnemyDiedEvent e2(ptr);
-    eb::EventBus::FireEvent(e2);
-  }
+  auto damage = hero->getDamage(shared_from_this());
+  applyDamage(hero, damage);
   return hp > 0;
+}
+
+void Enemy::onDamage(std::shared_ptr<Creature> attacker,
+                     std::shared_ptr<Damage> damage) {
+  auto ptr = shared_from_this();
+  EnemyTakeDamageEvent e(ptr, damage);
+  eb::EventBus::FireEvent(e);
+}
+
+void Enemy::onDie() {
+  auto ptr = shared_from_this();
+  passThrough = true;
+  EnemyDiedEvent e2(ptr);
+  eb::EventBus::FireEvent(e2);
 }
 
 Direction getDirFromCell(std::shared_ptr<Cell> c, Cell *nc) {